requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Data Collection
yfinance==0.2.28
praw==7.7.1
newsapi-python==0.2.7
tweepy==4.14.0

# Machine Learning & Deep Learning
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2

# Transformers & NLP
transformers==4.31.0
tokenizers==0.13.3
datasets==2.14.4
sentencepiece==0.1.99

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1

# Technical Indicators
ta==0.11.0
ta-lib==0.4.28

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Web Application
streamlit==1.25.0
gradio==3.41.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
jupyter==1.0.0
ipykernel==6.25.1

# API and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Model Evaluation
scikit-metrics==0.1.0
